Edits to test files
We have to source the R script that creates the object.
The objects are seen insede "test_that" functions, but
for some reasons the base package is given priority over these
in the search path, so the object called "labels" was masked by
the generic "labels". Now, it's called numLabels and it works.